+++ cmd/checkfleet/main.go
@@ -26,6 +26,7 @@ import (
 	"github.com/Allan-Nava/checkfleet/internal/checks/dns"
 	"github.com/Allan-Nava/checkfleet/internal/checks/haproxy"
 	"github.com/Allan-Nava/checkfleet/internal/checks/httpcheck"
+	"github.com/Allan-Nava/checkfleet/internal/checks/keycloak"
 	"github.com/Allan-Nava/checkfleet/internal/checks/nats"
 	"github.com/Allan-Nava/checkfleet/internal/checks/patroni"
 	"github.com/Allan-Nava/checkfleet/internal/checks/postgres"
@@ -74,7 +75,7 @@ func main() {
 
 func usage() {
 	fmt.Fprintln(os.Stderr, `uso:
-  checkfleet check <all|certs|http|nats|haproxy|stream|patroni|consul|postgres|dns|redis> --config checkfleet.yml [--output text|markdown|json|slack] [--only ...] [--min-severity warn] [--target glob] [--exit-on-bad]
+  checkfleet check <all|certs|http|nats|haproxy|stream|patroni|consul|postgres|dns|redis|keycloak> --config checkfleet.yml [--output text|markdown|json|slack] [--only ...] [--min-severity warn] [--target glob] [--exit-on-bad]
   checkfleet serve --config checkfleet.yml [--listen :9876] [--interval 60s]   # esporta le metriche Prometheus
   checkfleet report-issues --config checkfleet.yml [--dry-run]                 # apre/chiude issue GitHub dai finding BAD
   checkfleet validate --config checkfleet.yml                                  # valida la config senza eseguire i check
@@ -290,6 +291,7 @@ func modules(cfg *engine.Config) []moduleSpec {
 		{"postgres", c.Postgres != nil, func() engine.Check { return postgres.New(*c.Postgres) }},
 		{"dns", c.DNS != nil, func() engine.Check { return dns.New(*c.DNS) }},
 		{"redis", c.Redis != nil, func() engine.Check { return redis.New(*c.Redis) }},
+		{"keycloak", c.Keycloak != nil, func() engine.Check { return keycloak.New(*c.Keycloak) }},
 	}
 }